Adding Log-in and Register Pages
The log-in and register pages are included.
login.php: Contains features in the HTML where it is possible to
navigate back to the home page or the register page. Contains php
strings that allows a logged-in user to get to the welcome page and
have a possibility of viewing the secret page.
register.php: Contains features where it is possible to register an
account by filling in details (email, password, confirm password) or
locating back to the log-in page or home page. Contains php strings
that function as creating a new user and sending it to the database, in
strings $sql =.
All of the pages are styled using an external stylesheet, style.css.