Establishing a Connection
The database.php page is used for establishing a connection between the
server and the database. It is referred to in index.php, login.php and
register.php using “require 'database.php’;” tag as it needed for the
connecting of pages to server to database. The logout.php page is a
log-out feature that also directly navigates you back to the home page
(index.php).

index.php: Contains information that only logged-in users are able to
see using tags
<?php if( !empty($user) ): ?>
<?php else: ?>
<?php endif; ?>